vinculo pokemon com usuário

+++ lib/app/modules/my_application/src/authentication/external/services/pokemon_service.dart
@@ -1,5 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:login_with_firebase/models/poke_model.dart';
 
 class PokemonService {
   Future<void> savePokemon(String pokemonName) async {
@@ -8,10 +9,11 @@ class PokemonService {
       if (user != null) {
         final pokemonData = {
           'name': pokemonName,
-          'userId': user.uid,
         };
         await FirebaseFirestore.instance
-            .collection('Pokemons')
+            .collection('Users')
+            .doc(user.uid)
+            .collection('Pokemon')
             .add(pokemonData);
       }
     } catch (e) {
@@ -22,10 +24,15 @@ class PokemonService {
 
   Future<void> updatePokemonName(String pokemonId, String newName) async {
     try {
-      await FirebaseFirestore.instance
-          .collection('Pokemons')
-          .doc(pokemonId)
-          .update({'name': newName});
+      final user = FirebaseAuth.instance.currentUser;
+      if (user != null) {
+        await FirebaseFirestore.instance
+            .collection('Users')
+            .doc(user.uid)
+            .collection('Pokemon')
+            .doc(pokemonId)
+            .update({'name': newName});
+      }
     } catch (e) {
       print('Erro ao atualizar o nome do pokemon: $e');
       // Tratar o erro aqui se necessário.
@@ -34,10 +41,15 @@ class PokemonService {
 
   Future<void> deletePokemon(String pokemonId) async {
     try {
-      await FirebaseFirestore.instance
-          .collection('Pokemons')
-          .doc(pokemonId)
-          .delete();
+      final user = FirebaseAuth.instance.currentUser;
+      if (user != null) {
+        await FirebaseFirestore.instance
+            .collection('Users')
+            .doc(user.uid)
+            .collection('Pokemon')
+            .doc(pokemonId)
+            .delete();
+      }
     } catch (e) {
       print('Erro ao excluir o pokemon: $e');
       // Tratar o erro aqui se necessário.
